Change foreground color for highlighted text
Currently, highlighted text (for example, when using V to select lines) has a white foreground, and a light pink background.

This change makes the foreground black for more contrast, and it looks nice too! haha.

My main reasoning was that it was hard to see the highlighted text because the contrast between the white on the pink background wasn't enough.